Make item holder use the event itself

+++ src/main/java/com/saicone/picospacos/core/item/action/RestoreAction.java
@@ -3,6 +3,7 @@ package com.saicone.picospacos.core.item.action;
 import com.saicone.picospacos.api.item.ItemAction;
 import com.saicone.picospacos.api.item.ItemHolder;
 import com.saicone.picospacos.util.Items;
+import org.bukkit.event.entity.PlayerDeathEvent;
 import org.jetbrains.annotations.NotNull;
 
 public class RestoreAction implements ItemAction {
@@ -11,6 +12,9 @@ public class RestoreAction implements ItemAction {
 
     @Override
     public void execute(@NotNull ItemHolder holder) {
+        if (holder.getEvent() instanceof PlayerDeathEvent && ((PlayerDeathEvent) holder.getEvent()).getKeepInventory()) {
+            return;
+        }
         holder.getPlayerData().addItems(holder.getItem());
         holder.setItem(Items.empty());
     }
